:gear: tag official environmments by git revision
With {{cmake_re_source_hash}} the official environments are refered to with a tag matching the commit id of cmake-re, matching the version of cmake-re which is used to build the official environment, impacting the envzip hash with the cmake-re original git revision.

Therefore ensuring rebuilds of environments on new cmake-re releases.

This replaces the use of {{tipi_version_hash}} which referred to the released version of cmake-re.