Add a Ubuntu 24.04 image with a Musl sysroot for static linking
This creates a sysroot in /opt/sysroot/musl that contains:
- Musl v1.2.5
- Linux headers v6.17
- LLVM (Clang, LLD, runtimes, tools) v21.1.4

Clang will default to the target "x86_64-unknown-linux-musl" and
use the sysroot with all the right options for a correct runtime
using libc++, libc++abi, libunwind and compiler-rt.

It is required to link the binaries statically to avoid using the
Musl dynamic linker which is not to be distributed with the applications.